yumbook-backend/pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = ""
authors = ["Sagar Dwivedi <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = "^0.115.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
psycopg = { extras = ["binary"], version = "^3.2.2" }
sqlmodel = "^0.0.22"
alembic = "^1.13.2"
pyjwt = "^2.9.0"
pydantic-settings = "^2.5.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.6"
mypy = "^1.11.2"
types-passlib = "^1.7.7.20240819"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
preview = true
exclude = ["alembic"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions,
    "FAST",
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "W191", # indentation contains tabs
    "B904", # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
